+++ canisters/chain_fusion/ether/ether-rpc.service.ts
@@ -1,29 +1,61 @@
 import { RpcService, RpcServices } from "@bundly/ic-evm-rpc";
 
+export enum ChainIds {
+  EthSepolia = 11155111,
+  EthMainnet = 1,
+  ArbitrumOne = 42161,
+  BaseMainnet = 8453,
+  OptimismMainnet = 10,
+}
+
 export class EtherRpcService {
-  constructor(private service: RpcServices) {}
+  constructor(private services: RpcServices) {}
 
   public getChainId(): bigint {
-    return this.service.Custom ? this.service.Custom.chainId : 0n;
+    if (this.services.Custom) return this.services.Custom.chainId;
+    if (this.services.EthSepolia) return BigInt(ChainIds.EthSepolia);
+    if (this.services.EthMainnet) return BigInt(ChainIds.EthMainnet);
+    if (this.services.ArbitrumOne) return BigInt(ChainIds.ArbitrumOne);
+    if (this.services.BaseMainnet) return BigInt(ChainIds.BaseMainnet);
+    if (this.services.OptimismMainnet) return BigInt(ChainIds.OptimismMainnet);
+
+    throw new Error("ChainId not found");
   }
 
   public getValue(): RpcServices {
-    return this.service;
+    return this.services;
   }
 
   public getRequestService(): RpcService {
-    // TODO: add support for another services
-    if (!this.service.Custom?.services[0]) {
-      throw new Error("Service not found");
+    if (this.services.Custom) {
+      return {
+        Custom: {
+          url: this.services.Custom?.services[0].url,
+          headers: this.services.Custom?.services[0].headers,
+        },
+      };
     }
 
-    const service = {
-      Custom: {
-        url: this.service.Custom?.services[0].url,
-        headers: this.service.Custom?.services[0].headers,
-      },
-    };
+    if (this.services.EthSepolia && this.services.EthSepolia.Some?.[0]) {
+      return { EthSepolia: this.services.EthSepolia.Some?.[0] };
+    }
 
-    return service;
+    if (this.services.EthMainnet && this.services.EthMainnet.Some?.[0]) {
+      return { EthMainnet: this.services.EthMainnet.Some?.[0] };
+    }
+
+    if (this.services.ArbitrumOne && this.services.ArbitrumOne.Some?.[0]) {
+      return { ArbitrumOne: this.services.ArbitrumOne.Some?.[0] };
+    }
+
+    if (this.services.BaseMainnet && this.services.BaseMainnet.Some?.[0]) {
+      return { BaseMainnet: this.services.BaseMainnet.Some?.[0] };
+    }
+
+    if (this.services.OptimismMainnet && this.services.OptimismMainnet.Some?.[0]) {
+      return { OptimismMainnet: this.services.OptimismMainnet.Some?.[0] };
+    }
+
+    throw new Error("Service not found");
   }
 }