+++ canisters/chain_fusion/services/config.service.ts
@@ -1,16 +1,7 @@
-import {
-  None,
-  Opt,
-  Record,
-  Some,
-  StableBTreeMap,
-  ThresholdKeyInfo,
-  ecdsaPublicKey,
-  nat,
-  nat8,
-  text,
-} from "azle";
+import { None, Some, ThresholdKeyInfo, ecdsaPublicKey, nat } from "azle/experimental";
 import { computeAddress, hexlify } from "ethers";
+
+import { AppConfig, AppConfigStore } from "../database/database";
 
 export type EcdsaKeyIdName = "dfx_test_key" | "test_key_1" | "key_1";
 export type EcdsaKeyIdCurve = "secp256k1";
@@ -20,30 +11,17 @@ export type EcdsaKeyId = {
   name: EcdsaKeyIdName;
 };
 
-const AppConfig = Record({
-  ecdsaKeyId: Opt(
-    Record({
-      name: text,
-      curve: text,
-    })
-  ),
-  evmAddress: Opt(text),
-  nonce: nat,
-});
-type AppConfig = typeof AppConfig.tsType;
-
 export type InitOptions = {
   ecdsaKeyId: EcdsaKeyId;
 };
 
 export class ConfigService {
-  private configStore = StableBTreeMap<nat8, typeof AppConfig.tsType>(0);
   private config: AppConfig;
 
-  constructor() {
-    let appConfig = this.configStore.get(0).Some;
+  constructor(private readonly configStore: AppConfigStore) {
+    let appConfig = this.configStore.get(0);
 
-    if (appConfig === undefined) {
+    if (appConfig === null) {
       const newConfig = {
         ecdsaKeyId: None,
         evmAddress: None,