+++ canisters/chain_fusion/services/event.service.ts
@@ -1,34 +1,26 @@
-import { Opt, Record, StableBTreeMap, Vec, nat64, text } from "azle";
+import { nat64 } from "azle/experimental";
 
-import { RpcServices } from "@bundly/ic-evm-rpc";
-
-export const Event = Record({
-  service: RpcServices,
-  topics: Opt(Vec(Vec(text))),
-  addresses: Vec(text),
-  lastScrapedBlock: nat64,
-});
-export type Event = typeof Event.tsType;
-
-const events = StableBTreeMap<nat64, Event>(20);
+import { Event, EventStore } from "../database/database";
 
 export class EventService {
+  constructor(private readonly events: EventStore) {}
+
   public async add(data: Event) {
-    const nextId = events.len() + 1n;
-    events.insert(nextId, data);
+    const nextId = this.events.len() + 1n;
+    this.events.insert(nextId, data);
 
     return { id: nextId };
   }
 
   public getAll() {
-    return events.items();
+    return this.events.items();
   }
 
   public get(id: nat64) {
-    return events.get(id);
+    return this.events.get(id);
   }
 
   public update(id: nat64, contract: Event) {
-    return events.insert(id, contract);
+    return this.events.insert(id, contract);
   }
 }