+++ scripts/ic/stop.sh
@@ -1,18 +1,5 @@
 #!/bin/bash
 
-# Find process IDs listening on port 8545 (anvil)
-anvil=$(lsof -t -i:8545)
-
-# Check if any PIDs were found
-if [ -z "$anvil" ]; then
-    echo "Anvil not running."
-else
-    # Kill the processes
-    kill $anvil && echo "Terminated running Anvil process."
-fi
-
-# kill caddyserver
-caddy stop
 # kill dfx replica
 dfx stop
 # Find process IDs listening on port 4943 (dfx)